[Rebase] sync ipex-cpu related submodules. (#4014)
* sync ipex-cpu related submodules.

* sync csrc/cpu

* sync cmake/cpu

* sync test/cpu

* sync ipex/csrc/cpu frontend cpp code.

* build ipex cpu backend module pass.

* remove temprary disable cpu code.

+++ csrc/cpu/aten/RNN.cpp
@@ -318,7 +318,10 @@ std::vector<at::Tensor> lstm_kernel(
   }
 
   QuantizedLstmParams quantizedLstmParams(
-      {scale, zp, weights_scale_mask, weight_scales});
+      {(const float)scale,
+       (const int32_t)zp,
+       weights_scale_mask,
+       weight_scales});
   std::tie(w1_, w2_) = torch_ipex::cpu::get_lstm_packed_weight(
       weight_ih,
       weight_hh,
